requirements.txt
langgraph
langchain
langchain-community
langchain-groq
langchain-core
langchain-text-splitters
chromadb
pypdf
sentence-transformers
fastapi
uvicorn
streamlit
python-multipart
python-dotenv
groq
